refactor(backend): introduce report feature package
Moves the report controller, entity, repository and DTOs out of the
generic controller/database/dto layer packages into a new
net.onelitefeather.blackhole.backend.report package, its own feature
package rather than folded into punishment despite calling
PunishmentApplicationService and EloService. Pure physical move, no
behavior change.
