+++ backend/src/main/java/net/onelitefeather/blackhole/backend/report/ReportRepository.java
@@ -1,8 +1,7 @@
-package net.onelitefeather.blackhole.backend.database.repository;
+package net.onelitefeather.blackhole.backend.report;
 
 import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.repository.PageableRepository;
-import net.onelitefeather.blackhole.backend.database.entities.ReportEntity;
 
 import java.util.UUID;
