+++ backend/src/main/java/net/onelitefeather/blackhole/backend/report/ReportResolutionDTO.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.dto;
+package net.onelitefeather.blackhole.backend.report;
 
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.annotation.Nullable;
